Add pawn structure evaluation
Identify isolated, doubled, backward, and passed pawns in addition to
pawns that have a neighbor on the adjacent eastern file and defended
pawns. Store these evaluations in a hash table to prevent costly
re-evaluation.

+++ src/transposition_table.cc
@@ -5,15 +5,16 @@
  * Licensed under MIT License. Terms and conditions enclosed in "LICENSE.txt".
  */
 
+#include "transposition_table.h"
+
 #include "board.h"
 #include "move.h"
-#include "transposition_table.h"
 
 namespace omegazero {
 
 constexpr U64 kHashMask = 0X7FFFF;
 
-auto TranspositionTable::Access(const Board* board, int depth, int& eval,
+auto TranspositionTable::Access(const Board* board, int depth, float& eval,
                                 S8& node_type) const -> bool {
   U64 board_hash = board->GetBoardHash();
   int index = board_hash & kHashMask;
@@ -82,7 +83,7 @@ auto TranspositionTable::GetHashMove(const Board* board) const -> Move {
   return hash_move;
 }
 
-auto TranspositionTable::Update(const Board* board, int depth, int eval,
+auto TranspositionTable::Update(const Board* board, int depth, float eval,
                                 S8 node_type, const Move& hash_move) -> void {
   TableEntry new_entry;
   new_entry.hash_move = hash_move;
